requirements.txt
Pillow
tqdm
pyyaml
colorama